+++ crate/sprite_loading/src/sprite_sheet_mapper.rs
@@ -223,7 +223,7 @@ mod test {
         }
     }
 
-    fn offsets(n: usize) -> Vec<SpriteOffset> {
-        (0..n).map(|_| (0, 0).into()).collect()
+    fn offsets(n: usize) -> Option<Vec<SpriteOffset>> {
+        Some((0..n).map(|_| (0, 0).into()).collect())
     }
 }